chore: move total surplus fetching to updater and store it on jotai state
To avoid re-fetching every time the account activity modal is open
