+++ src/modules/application/containers/App/Updaters.tsx
@@ -5,23 +5,24 @@ import GasUpdater from 'legacy/state/gas/updater'
 import ListsUpdater from 'legacy/state/lists/updater'
 import LogsUpdater from 'legacy/state/logs/updater'
 import {
-  GpOrdersUpdater,
   CancelledOrdersUpdater,
+  ExpiredOrdersUpdater,
+  GpOrdersUpdater,
   PendingOrdersUpdater,
   UnfillableOrdersUpdater,
-  ExpiredOrdersUpdater,
 } from 'legacy/state/orders/updaters'
 import { SpotPricesUpdater } from 'legacy/state/orders/updaters/SpotPricesUpdater'
 import FeesUpdater from 'legacy/state/price/updater'
 import SentryUpdater from 'legacy/state/sentry/updater'
 import UserUpdater from 'legacy/state/user/updater'
 
-import { UploadToIpfsUpdater, AppDataUpdater } from 'modules/appData'
+import { AppDataUpdater, UploadToIpfsUpdater } from 'modules/appData'
 import { InjectedWidgetUpdater } from 'modules/injectedWidget'
-import { EthFlowSlippageUpdater, EthFlowDeadlineUpdater } from 'modules/swap/state/EthFlow/updaters'
+import { EthFlowDeadlineUpdater, EthFlowSlippageUpdater } from 'modules/swap/state/EthFlow/updaters'
 import { TokensListUpdater } from 'modules/tokensList/updaters/TokensListUpdater'
 import { WalletUpdater } from 'modules/wallet'
 
+import { TotalSurplusUpdater } from 'common/state/totalSurplusState'
 import { ThemeFromUrlUpdater } from 'common/updaters/ThemeFromUrlUpdater'
 import { MulticallUpdater } from 'lib/state/multicall'
 
@@ -52,6 +53,7 @@ export function Updaters() {
       <SpotPricesUpdater />
       <ThemeFromUrlUpdater />
       <InjectedWidgetUpdater />
+      <TotalSurplusUpdater />
     </>
   )
 }
